🚚 Rename: 룰렛 버튼 폴더 구조 변경
