+++ src/components/mainPage/groupRouletteButton/groupRouletteButton.module.scss
